requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.12.2
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
loguru==0.5.3
MarkupSafe==1.1.1
packaging==20.7
paho-mqtt==1.5.1
psycopg2==2.7.5
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
